requirements.txt
setuptools~=72.1.0
fastapi~=0.112.0
uvicorn~=0.30.5
python-dotenv~=1.0.1
librosa~=0.10.2
music21~=9.1.0